refactor: 메시지 아이템 분리하여 메시지 컨테이너 계층 변경
